S12 p11 a208 304 be tag controller 설치하기

See merge request s12-webmobile2-sub1/S12P11A208!83

+++ src/main/java/com/be/byeoldam/domain/common/repository/TagBookmarkUrlRepositoryCustom.java
@@ -5,5 +5,5 @@ import com.be.byeoldam.domain.tag.dto.RecommendedUrlResponse;
 import java.util.List;
 
 public interface TagBookmarkUrlRepositoryCustom {
-    List<RecommendedUrlResponse> findBookmarkUrlsByTagName(String tagName, Long cursorId, int size);
+    List<RecommendedUrlResponse> findBookmarkUrlsByTagName(String tagName, int cursorId, int size);
 }
